支持系列内手动编辑排序序号（Komga 式锁定），并完善步进器交互。

+++ app/test/domain/models/entity/comic/series_volume_labels_test.dart
@@ -13,7 +13,7 @@ Series _series({int itemCount = 3, int? totalCount}) {
     totalCount: totalCount,
     items: List<SeriesItem>.generate(
       itemCount,
-      (int index) => SeriesItem(comicId: 'comic-$index', order: index),
+      (int index) => SeriesItem(comicId: 'comic-$index', order: index.toDouble()),
     ),
   );
 }
